Created and properly implemented the Enemy Manager class that manages enemies and launches them on a specific interval.
Currently, PlayState tells the Enemy Manager to create a pool of 100 enemies. This should be sufficient for testing on lower end machines, as I can't imagine having an area of the game with more than 100 enemies.

I also fixed a small bug with how bullets expired (previously if you shot through the allocated bullets, any new ones would be automatically killed). It is working properly now.